Moved retry session to giosg.py
