+++ server/giosg.py
@@ -1,5 +1,7 @@
 from conf import SERVICE_URL
-from retry_session import retry_session
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
+import requests
 
 
 class APIClient:
@@ -54,3 +56,21 @@ class APIClient:
             for result in page['results']:
                 yield result
             url = page['next']
+
+
+def retry_session(retries, session=None, backoff_factor=0.3, status_forcelist=(500, 502, 503, 504)):
+    """
+    Custom implementation of requests to enable retry policy
+    """
+    session = session or requests.Session()
+    retry = Retry(
+        total=retries,
+        read=retries,
+        connect=retries,
+        backoff_factor=backoff_factor,
+        status_forcelist=status_forcelist,
+    )
+    adapter = HTTPAdapter(max_retries=retry)
+    session.mount('http://', adapter)
+    session.mount('https://', adapter)
+    return session
