--- server/retry_session.py
@@ -1,21 +1,0 @@
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
-import requests
-
-
-def retry_session(retries, session=None, backoff_factor=0.3, status_forcelist=(500, 502, 503, 504)):
-    """
-    Custom implementation of requests to enable retry policy
-    """
-    session = session or requests.Session()
-    retry = Retry(
-        total=retries,
-        read=retries,
-        connect=retries,
-        backoff_factor=backoff_factor,
-        status_forcelist=status_forcelist,
-    )
-    adapter = HTTPAdapter(max_retries=retry)
-    session.mount('http://', adapter)
-    session.mount('https://', adapter)
-    return session
